youtube_auto/requirements.txt
openai>=1.30.0
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.1
moviepy>=1.0.3
Pillow>=10.0.0
requests>=2.31.0
